Add files via upload
Objective of study is to develop model to predict median value of homes in the Boston area using data collected in the 1970s. 

Diagram uses a variety of regression models and decision trees and evaluates for optimization of end model to make best predictions of housing prices.
End model selected was a gradient boost model that iterates 100 times, with ASE of 13.144 on the validation data set.