Add files via upload
Project objective to classify whether an email is junk email (1) or not junk (0). Machine Learning models used to find the optimal model are a variety of decision trees, Logistic Regression, and an Ensemble model. Among the models, the logistic regression model performs the highest, with a misclassification rate of 7.45% of the validation data. Some parameters (words) that have the most significant p-values on determining whether or not a particular observation is a spam email include: "Free","Remove", "Business".